Move BoardColor class to separate module
